backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pytest==7.4.3
httpx==0.25.1
google-generativeai==0.3.1
